+++ src/scales/scale.linearbase.js
@@ -1,6 +1,7 @@
 'use strict';
 
 import helpers from '../helpers/index';
+import {almostEquals, almostWhole, _decimalPlaces, _setMinAndMaxByKey, sign} from '../helpers/helpers.math';
 import Scale from '../core/core.scale';
 
 const isNullOrUndef = helpers.isNullOrUndef;
@@ -43,7 +44,7 @@ function generateTicks(generationOptions, dataRange) {
 
 	if (stepSize || isNullOrUndef(precision)) {
 		// If a precision is not specified, calculate factor based on spacing
-		factor = Math.pow(10, helpers._decimalPlaces(spacing));
+		factor = Math.pow(10, _decimalPlaces(spacing));
 	} else {
 		// If the user specified a precision, round to that number of decimal places
 		factor = Math.pow(10, precision);
@@ -56,17 +57,17 @@ function generateTicks(generationOptions, dataRange) {
 	// If min, max and stepSize is set and they make an evenly spaced scale use it.
 	if (stepSize) {
 		// If very close to our whole number, use it.
-		if (!isNullOrUndef(min) && helpers.almostWhole(min / spacing, spacing / 1000)) {
+		if (!isNullOrUndef(min) && almostWhole(min / spacing, spacing / 1000)) {
 			niceMin = min;
 		}
-		if (!isNullOrUndef(max) && helpers.almostWhole(max / spacing, spacing / 1000)) {
+		if (!isNullOrUndef(max) && almostWhole(max / spacing, spacing / 1000)) {
 			niceMax = max;
 		}
 	}
 
 	numSpaces = (niceMax - niceMin) / spacing;
 	// If very close to our rounded value, use it.
-	if (helpers.almostEquals(numSpaces, Math.round(numSpaces), spacing / 1000)) {
+	if (almostEquals(numSpaces, Math.round(numSpaces), spacing / 1000)) {
 		numSpaces = Math.round(numSpaces);
 	} else {
 		numSpaces = Math.ceil(numSpaces);
@@ -103,8 +104,8 @@ class LinearScaleBase extends Scale {
 		// do nothing since that would make the chart weird. If the user really wants a weird chart
 		// axis, they can manually override it
 		if (opts.beginAtZero) {
-			var minSign = helpers.sign(me.min);
-			var maxSign = helpers.sign(me.max);
+			var minSign = sign(me.min);
+			var maxSign = sign(me.max);
 
 			if (minSign < 0 && maxSign < 0) {
 				// move the top up to 0
@@ -215,7 +216,7 @@ class LinearScaleBase extends Scale {
 
 		// At this point, we need to update our max and min given the tick values since we have expanded the
 		// range of the scale
-		helpers._setMinAndMaxByKey(ticks, me, 'value');
+		_setMinAndMaxByKey(ticks, me, 'value');
 
 		if (opts.reverse) {
 			ticks.reverse();
