+++ src/scales/scale.radialLinear.js
@@ -2,6 +2,7 @@
 
 import defaults from '../core/core.defaults';
 import helpers from '../helpers/index';
+import {isNumber, toDegrees} from '../helpers/helpers.math';
 import LinearScaleBase from './scale.linearbase';
 import Ticks from '../core/core.ticks';
 
@@ -156,7 +157,7 @@ function fitWithPointLabels(scale) {
 
 		// Add quarter circle to make degree 0 mean top of circle
 		var angleRadians = scale.getIndexAngle(i);
-		var angle = helpers.toDegrees(angleRadians) % 360;
+		var angle = toDegrees(angleRadians) % 360;
 		var hLimits = determineLimits(angle, pointPosition.x, textSize.w, 0, 180);
 		var vLimits = determineLimits(angle, pointPosition.y, textSize.h, 90, 270);
 
@@ -239,7 +240,7 @@ function drawPointLabels(scale) {
 		ctx.fillStyle = pointLabelFontColor;
 
 		var angleRadians = scale.getIndexAngle(i);
-		var angle = helpers.toDegrees(angleRadians);
+		var angle = toDegrees(angleRadians);
 		ctx.textAlign = getTextAlignForAngle(angle);
 		adjustPointPositionForLabelHeight(angle, scale._pointLabelSizes[i], pointLabelPosition);
 		fillText(ctx, scale.pointLabels[i], pointLabelPosition, plFont.lineHeight);
@@ -287,7 +288,7 @@ function drawRadiusLine(scale, gridLineOpts, radius, index) {
 }
 
 function numberOrZero(param) {
-	return helpers.isNumber(param) ? param : 0;
+	return isNumber(param) ? param : 0;
 }
 
 class RadialLinearScale extends LinearScaleBase {
